requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
langchain==1.2.13
openai==2.30.0
python-jose==3.5.0
slowapi==0.1.9
langchain-openai==1.1.12
python-dotenv==1.2.2
python-multipart==0.0.22
llama-index==0.14.20
llama-index-llms-azure-openai==0.5.3
llama-index-embeddings-azure-openai==0.5.2
